requirements.txt
# Core AI/ML libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
diffusers>=0.21.0
transformers>=4.25.0
accelerate>=0.20.0
peft>=0.5.0
safetensors>=0.3.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Data handling
numpy>=1.21.0
datasets>=2.10.0
requests>=2.28.0

# Web interface
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# Utilities
tqdm>=4.64.0
matplotlib>=3.5.0
scipy>=1.9.0

# Optional: For better performance
xformers>=0.0.20; sys_platform != "darwin"

# Development tools (optional)
jupyter>=1.0.0
ipywidgets>=8.0.0